I think I've done an alright job compressing like-terms in polynomial expressions. Definitely need to do more testing, but so far it seems to be doing the right thing. :)
...Well, other than one hold up: I currently have two functions (compress*
and compress+) which have conditions that track constants through hash
tables (this is how we're matching like-terms). However, if a key gets
zeroed out (e.g., we see both "x" and "-x" during compress+), the value
doesn't get culled, so you'll get an output like "(* 0 x)". Not the end of
the world, but obviously not what we want long term.